Naming convention, second of several.
More changes to reflect the updated naming convention. This set of
changes relates to modules in mergeron.core, and completes module name
changes to reflect the updated naming convention. We now turn to
prefixes to function names and variable names, which may add verbiage
without clarifying the code, as implied by PEP8.