feature: provide helper for for UPP boundaries
Provide a helper function for generatin UPP boundaries using various
aggregators, and let the helper call the aggregator-specific functions
based on attribute values in a UPPBoundarySpec object; all parameter
validation in the individual functions in moved to the constructor for
the UPPBoundarySpec object; rename some parameters for convenience,
move specialized functions for generating UPP boundaries to
a separate module.

BREAKING CHANGE: In functions for generating UPP boundaries,
"avg_method" was renamed to "agg_method" and "wgtng_policy" was renamed
to "weighting".